Fixed compile on Windows in release mode.
It is not valid to unconditionally enable debug mode on platforms where
debug/release affects linkage.  On these platforms, debug mode can only
be used if the debug libraries are enabled, and release mode can only be
used if the release libraries are enabled.

Let the build system set the correct mode.

Change-Id: I1a4869a47b158049de2e8c71a62c3df059a70281