Update the git-archive export options
The .gitattributes and .gitignore files do not need to be present in
packaged sources, as they are for people using the Git repository.
People who download tarballs usually don't reimport into Git -- they
should just download the actual repositories.

Meanwhile, the .tag file is quite useful. Instead of making Git extract
to the useless SHA-1 of the $n string, ask git-archive to store
the SHA-1 of the commit being packaged.

Change-Id: Ie706262d09be3274da37c085a5adbeac7b3df7ac
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@nokia.com>